frontend/requirements.txt
pandas==1.4.2
streamlit==1.13.0
requests==2.27.1
matplotlib==3.5.1
seaborn==0.11.2
joblib==1.1.0
optuna==3.0.2
PyYAML==6.0
python-multipart==0.0.5
plotly==5.6.0
scikit-learn==1.0.2
lightgbm==3.2.1